+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Imp/EnactSpec.hs
@@ -1,4 +1,3 @@
-{-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE OverloadedLists #-}
 {-# LANGUAGE ScopedTypeVariables #-}
@@ -14,16 +13,19 @@ import Cardano.Ledger.Conway.Governance
 import Cardano.Ledger.Conway.Rules (EnactSignal (..))
 import Cardano.Ledger.Shelley.LedgerState
 import Lens.Micro.Mtl
-import Test.Cardano.Ledger.Common
 import Test.Cardano.Ledger.Conway.ImpTest
+import Test.Cardano.Ledger.Imp.Common
 
-spec :: forall era. (ConwayEraImp era, GovState era ~ ConwayGovState era) => Spec
+spec ::
+  forall era.
+  (ConwayEraImp era, GovState era ~ ConwayGovState era) =>
+  SpecWith (ImpTestState era)
 spec =
   describe "ENACT" $ do
-    itM @era "TreasuryWithdrawal" $ do
+    it "TreasuryWithdrawal" $ do
       rewardAcount1 <- registerRewardAccount
       let govAction = TreasuryWithdrawals [(rewardAcount1, Coin 666)]
-      govActionId <- submitProposal govAction
+      govActionId <- submitGovAction govAction
       enactStateInit <- use $ impNESL . newEpochStateGovStateL . cgEnactStateL
       let signal =
             EnactSignal
@@ -35,7 +37,7 @@ spec =
               { ensTreasury = Coin 1000
               }
       enactState' <- runImpRule @"ENACT" () enactState signal
-      impIO $ ensWithdrawals enactState' `shouldBe` [(getRwdCred rewardAcount1, Coin 666)]
+      ensWithdrawals enactState' `shouldBe` [(getRwdCred rewardAcount1, Coin 666)]
 
       rewardAcount2 <- registerRewardAccount
       let withdrawals' =
@@ -43,7 +45,7 @@ spec =
             , (rewardAcount2, Coin 222)
             ]
           govAction' = TreasuryWithdrawals withdrawals'
-      govActionId' <- submitProposal govAction'
+      govActionId' <- submitGovAction govAction'
       let signal' =
             EnactSignal
               { esGovActionId = govActionId'
@@ -52,9 +54,8 @@ spec =
 
       enactState'' <- runImpRule @"ENACT" () enactState' signal'
 
-      impIO $ do
-        ensWithdrawals enactState''
-          `shouldBe` [ (getRwdCred rewardAcount1, Coin 777)
-                     , (getRwdCred rewardAcount2, Coin 222)
-                     ]
-        ensTreasury enactState'' `shouldBe` Coin 1
+      ensWithdrawals enactState''
+        `shouldBe` [ (getRwdCred rewardAcount1, Coin 777)
+                   , (getRwdCred rewardAcount2, Coin 222)
+                   ]
+      ensTreasury enactState'' `shouldBe` Coin 1